Fixes the following issues (mainly for Samsung S3 on 4.1.2):
* On Samsung S3: reading an AAC file failed. A raw AAC stream starts with
2 bytes giving basic info about the stream (e.g. sampling frequency,
channel configuration). However if fed to the S3 decoder, MediaCodec will
crash. So now these two bytes are not fed to the encoder (MediaCodec has
already been configured and knows these parameters). This looks OK with
other devices too (e.g. Nexus 5 and 7 running Lollipop).
* On Samsung S3: AAC mono files are not decoded properly. The decoder
returns duplicated samples, as if the stream was stereo (maybe because
of the missing first 2 bytes...). This results is an audio stream with
a much lower pitch, containing half the data from the original stream
(Cf. length of the decoded stream is still the same as the original, not
twice). To solve that, the encoder now always encodes at least in stereo.
* Some device seem to fail encoding in AAC. I couldn't replicate the
issue with either my devices or several emulators. So now, until the root
cause is found, if the AAC encoding fails then the samples will be
exported in a WAVE file instead. (files will be much bigger than the .m4a
files, but ringtones are usually fairly small so it's should be fine).
* Make the AAC stream 2 frames (2048 samples) bigger than the original
stream: the encoder adds 2 empty frames (priming frames) before
processing the real data. As a result, the decoded samples contain
2 empty frames at the begining and are missing the last two frames. By
making the AAC file 2 frames bigger, the decoded samples will still
contain the 2 empty frames at the begining, but will at least contain
the end of the stream.